# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Generation Service",
    "description": "Core services for template processing and LLM interaction.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/store/presign": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Generate Upload Urls",
        "description": "Generate presigned PUT URLs. Does NOT create DB records yet.",
        "operationId": "generateFileUploadUrls",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/FilePresignRequest"
                },
                "type": "array",
                "title": "Files"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FilePresignResponse"
                  },
                  "type": "array",
                  "title": "Response Generatefileuploadurls"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/single": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Create File",
        "description": "Standard Create Endpoint.\nClient calls this AFTER successfully uploading to the presigned URL.\nAutomatically generates PDF preview for DOCX/XLSX files.",
        "operationId": "createFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Create Files",
        "description": "Create Endpoint for multiple files.\nAutomatically generates PDF previews for DOCX/XLSX files.",
        "operationId": "createFiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FileCreate"
                },
                "title": "Files In"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileRead"
                  },
                  "title": "Response Createfiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Read Files",
        "description": "Retrieve files.",
        "operationId": "readFiles",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileRead"
                  },
                  "title": "Response Readfiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/{file_id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Read File",
        "description": "Get file metadata by ID.",
        "operationId": "readFile",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Storage"
        ],
        "summary": "Update File",
        "description": "Update file metadata (e.g., attach to job/project).",
        "operationId": "updateFile",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/{file_id}/download": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Generate Download Url",
        "description": "Generate a presigned download URL for a file.\n\nArgs:\n    inline: If True, sets content-disposition to inline (for browser preview).\n           If False (default), sets to attachment (for download).",
        "operationId": "generateFileDownloadUrl",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          },
          {
            "name": "inline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Inline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Generatefiledownloadurl"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/instructions/{instruction_id}/files": {
      "get": {
        "tags": [
          "Instructions"
        ],
        "summary": "Read Instruction Files",
        "description": "Get all files attached to an instruction.",
        "operationId": "readInstructionFiles",
        "parameters": [
          {
            "name": "instruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileRead"
                  },
                  "title": "Response Readinstructionfiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/surveys/{survey_id}/files": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Read Survey Files",
        "description": "Get all files attached to a survey.",
        "operationId": "readSurveyFiles",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileRead"
                  },
                  "title": "Response Readsurveyfiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileCreate": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 512,
            "title": "File Name"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 255,
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "org_id": {
            "type": "integer",
            "title": "Org Id"
          },
          "uploaded_by_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By User Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "instruction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction Id"
          },
          "survey_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survey Id"
          },
          "role": {
            "$ref": "#/components/schemas/FileRole",
            "default": "input"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "mime_type",
          "storage_key",
          "org_id"
        ],
        "title": "FileCreate"
      },
      "FilePresignRequest": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 512,
            "title": "File Name"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 255,
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "mime_type",
          "client_id"
        ],
        "title": "FilePresignRequest"
      },
      "FilePresignResponse": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 512,
            "title": "File Name"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 255,
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "put_url": {
            "type": "string",
            "title": "Put Url"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "mime_type",
          "client_id",
          "put_url",
          "storage_key"
        ],
        "title": "FilePresignResponse"
      },
      "FileRead": {
        "properties": {
          "file_name": {
            "type": "string",
            "maxLength": 512,
            "title": "File Name"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 255,
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "uploaded_by_user_id": {
            "type": "integer",
            "title": "Uploaded By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role": {
            "$ref": "#/components/schemas/FileRole"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "instruction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction Id"
          },
          "survey_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survey Id"
          },
          "preview_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview File Id"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "mime_type",
          "id",
          "storage_key",
          "uploaded_by_user_id",
          "created_at",
          "role"
        ],
        "title": "FileRead"
      },
      "FileRole": {
        "type": "string",
        "enum": [
          "template",
          "preview_pdf",
          "artefact",
          "input"
        ],
        "title": "FileRole"
      },
      "FileUpdate": {
        "properties": {
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview File Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "instruction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction Id"
          },
          "survey_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survey Id"
          }
        },
        "type": "object",
        "title": "FileUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}